refactor name utils to notifications since it describes it better
