+++ ERP/ERP/urls.py
@@ -24,5 +24,5 @@ urlpatterns = [
     path('manufacturing/', include('manufacturing.urls')),
     path('accounting/', include('accounting.urls')),
     path('sales/', include('sales.urls')),
-    path('subscriptions/', include('utils.urls'))
+    path('subscriptions/', include('notifications.urls'))
 ]
