--- ERP/utils/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = 'utils.apps.UtilsConfig'
